Move codes.py into core
