+++ grid/app/main/events/sfl/control_events.py
@@ -2,7 +2,7 @@
 import json
 
 # Local imports
-from ...codes import MSG_FIELD
+from ...core.codes import MSG_FIELD
 
 
 def socket_ping(message: dict, socket) -> str:
