+++ grid/app/main/routes/sfl/routes.py
@@ -23,7 +23,7 @@ from ...sfl.models import model_manager
 from ...sfl.workers import worker_manager
 from ...sfl.processes import process_manager
 from ...sfl.syft_assets import plans, protocols
-from ...codes import RESPONSE_MSG, CYCLE, MSG_FIELD
+from ...core.codes import RESPONSE_MSG, CYCLE, MSG_FIELD
 from ...events.sfl.fl_events import report, cycle_request, assign_worker_id
 from ...sfl.auth.federated import verify_token
 from ...exceptions import InvalidRequestKeyError, PyGridError
