+++ grid/app/main/sfl/auth/federated.py
@@ -1,5 +1,5 @@
 # PyGrid imports
-from ...codes import MSG_FIELD, RESPONSE_MSG, CYCLE, FL_EVENTS
+from ...core.codes import RESPONSE_MSG
 from ..processes import process_manager
 
 # Generic imports
